feat: implement ReactFinalForm on ext. data comp.
